spoolapp/requirements.txt
asgiref==3.3.1
Django==3.1.6
django-cors-headers==3.7.0
django-registration-redux==2.9
django-tailwind==1.1.1
djangorestframework==3.12.2
pytz==2021.1
sqlparse==0.4.1
python-dotenv
python-dotenv
asgiref==3.3.1
Django==3.1.6
django-cors-headers==3.7.0
django-lockdown==4.0.0
django-registration-redux==2.9
django-tailwind==1.1.1
djangorestframework==3.12.2
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.4.1
